requirements.txt
torch
torchvision
pandas
numpy
Pillow
scikit-learn
matplotlib
seaborn
tqdm

